Curency denomination distibutor original/packages/local-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiosqlite==0.19.0
python-dateutil==2.8.2

# PDF generation
reportlab==4.0.7

# Excel generation
openpyxl==3.1.2

# CSV
# Built-in Python csv module

# OCR and Document Processing
pytesseract==0.3.10        # Tesseract OCR wrapper
Pillow==10.1.0             # Image processing
PyMuPDF==1.23.8            # PDF text extraction (fitz)
pdf2image==1.16.3          # PDF to image conversion
python-docx==1.1.0         # Word document processing
opencv-python==4.8.1.78    # Image preprocessing (optional)
